requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
absl-py==1.4.0
accelerate==0.21.0
albumentations==1.3.1
braceexpand==0.1.7
datasets==2.14.0
deepspeed==0.10.0
diffusers==0.20.0
einops==0.6.1
fairscale==0.4.13
flash-attn==2.0.4
mmcv-full==1.7.0
ninja==1.11.1
nltk==3.8.1
Pillow==10.0.0
pyarrow==12.0.1
pycocoevalcap==1.2
pycocotools==2.0.6
scikit-learn==1.3.1
scipy==1.11.1
sentencepiece==0.1.99
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1+cu117
transformers==4.31.0
triton==2.0.0
webdataset==0.2.48
xformers==0.0.20
omegaconf==2.3.0
peft==0.3.0
